PERA-3127 Use double instead of float for slippage tolerance

+++ app/src/main/kotlin/com/algorand/android/ui/swap/view/SwapScreenContentState.kt
@@ -127,7 +127,7 @@ fun SwapScreenContentState(
                 onApplyClick = {
                     swapViewModel.applySwapConfigs(it)
                     if (it.balancePercentage != null) {
-                        val percentage = it.balancePercentage.toInt()
+                        val percentage = it.balancePercentage
                         widgetViewModel.setAmountByPercentage(swapViewModel.getSwapDetails(), percentage)
                     }
                     swapConfigurationBottomSheetState.hide()
